doc: Explain that escaped quotes require quoted queries
This change also refactors all the
fixture-like test code into conftest.py.
